requirements_eb-man.txt
absl-py==2.1.0
accelerate==1.2.1
addict==2.4.0
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
amsolver==1.1.0
annotated-types==0.7.0
anthropic==0.45.0
antlr4-python3-runtime==4.9.3
anyio==4.7.0
async-timeout==5.0.1
av==14.0.1
beautifulsoup4==4.12.3
blinker==1.9.0
cachetools==5.5.1
cffi==1.14.2
charset-normalizer==3.4.0
click==8.1.8
cloudpickle==3.1.0
ConfigArgParse==1.7
contourpy==1.3.0
dash==2.18.2
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
datasets==3.2.0
defusedxml==0.7.1
dill==0.3.8
diskcache==5.6.3
distro==1.9.0
einops==0.8.0
Farama-Notifications==0.0.4
fastapi==0.115.6
fastjsonschema==2.21.1
filelock==3.16.1
fire==0.7.0
Flask==3.0.3
fonttools==4.55.3
frozenlist==1.5.0
fsspec==2024.9.0
gdown==5.2.0
google-ai-generativelanguage==0.6.15
google-api-core==2.24.0
google-api-python-client==2.159.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
google-generativeai==0.8.4
googleapis-common-protos==1.66.0
grpcio==1.70.0
grpcio-status==1.70.0
gym-notices==0.0.8
gymnasium==1.0.0
h11==0.14.0
html-testRunner==1.2.1
httpcore==1.0.7
httplib2==0.22.0
httpx==0.28.1
huggingface-hub==0.27.0
hydra-core==1.3.2
idna==3.10
ImageIO==2.37.2
importlib_resources==6.5.2
interegular==0.3.3
ipywidgets==8.1.5
itsdangerous==2.2.0
Jinja2==3.1.5
jiter==0.8.2
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyterlab_widgets==3.0.13
lark==1.2.2
llvmlite==0.46.0
lmdeploy==0.12.1
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.9.4
mdurl==0.1.2
mmengine-lite==0.10.5
mpmath==1.3.0
msgpack==1.1.2
multidict==6.1.0
multiprocess==0.70.16
natsort==8.4.0
nbformat==5.10.4
networkx==3.2.1
numpy==1.26.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-ml-py==12.560.30
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.1.105
omegaconf==2.3.0
open3d==0.18.0
openai==1.58.1
openai-harmony==0.0.8
opencv-python==4.11.0.86
opencv-python-headless==4.11.0.86
outlines==0.0.46
pandas==2.2.3
partial-json-parser==0.2.1.1.post7
peft==0.11.1
pillow==11.0.0
plotly==5.24.1
prometheus_client==0.24.1
propcache==0.2.1
proto-plus==1.25.0
protobuf==5.29.2
py-cpuinfo==9.0.0
pyarrow==18.1.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pybase64==1.4.3
pycountry==24.6.1
pycparser==2.22
pydantic==2.12.5
pydantic_core==2.41.5
pynvml==12.0.0
pyquaternion==0.9.9
pytz==2024.2
PyYAML==6.0.2
qwen-vl-utils==0.0.8
ray==2.54.0
referencing==0.35.1
regex==2024.11.6
requests==2.32.3
retrying==1.3.4
rich==13.9.4
rpds-py==0.22.3
rsa==4.9
safetensors==0.4.5
scikit-learn==1.6.0
scipy==1.13.1
seaborn==0.13.2
segment-anything==1.0
sentencepiece==0.2.0
shortuuid==1.0.13
sniffio==1.3.1
soupsieve==2.6
stable_baselines3==2.4.0
starlette==0.41.3
supervision==0.25.1
sympy==1.13.1
tenacity==9.0.0
termcolor==2.5.0
threadpoolctl==3.5.0
tiktoken==0.8.0
tokenizers==0.21.0
tomli==2.2.1
torch==2.4.0
torchvision==0.19.0
transform3d==0.0.4
transformers==4.49.0
transforms3d==0.4.2
triton==3.0.0
typing-inspection==0.4.2
tzdata==2024.2
ultralytics==8.3.61
ultralytics-thop==2.0.13
uritemplate==4.1.1
urllib3==2.2.3
uvicorn==0.34.0
Werkzeug==3.0.6
widgetsnbextension==4.0.13
xgrammar==0.1.31
xxhash==3.5.0
yapf==0.43.0
yarl==1.18.3
